requirements.txt
Flask==2.1.0
python-dotenv==0.19.1
openai
pdfkit==0.6.1
weasyprint